refactor : clean up unnessernary code and update pagination.
